# contracts/team-core.openapi.yaml
openapi: 3.1.0
info:
  title: Smart Campus — Core Business API Contract
  version: 1.0.0
  description: |
    Hợp đồng API (OpenAPI 3.1) cho Core Business (Nhóm B6).
    Đã định nghĩa sẵn các Endpoint nhận Event từ Access Gate (B3) và IoT (B1), cũng như xử lý Cảnh báo (Alerts).
  contact:
    name: FIT4110 Teaching Team
    email: fit4110@example.edu.vn
servers:
  - url: http://localhost:4010
    description: Prism mock server
  - url: https://api.campus.local
    description: Class network placeholder
tags:
  - name: health
    description: Kiểm tra trạng thái service
  - name: alerts
    description: Quản lý cảnh báo Smart Campus
  - name: events
    description: Quản lý sự kiện nghiệp vụ
  - name: access
    description: Tích hợp Access Gate
  - name: vision
    description: Tích hợp AI Vision
security:
  - bearerAuth: []
paths:
  /health:
    get:
      operationId: getHealth
      summary: Kiểm tra trạng thái service
      description: Endpoint kiểm tra service còn hoạt động.
      tags: [health]
      security: []
      responses:
        '200':
          description: Service đang hoạt động
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
              examples:
                ok:
                  value:
                    status: ok
                    service: lab02-contract-template
                    time: '2026-05-10T08:00:00Z'

  /alerts:
    post:
      operationId: createAlert
      summary: Tạo cảnh báo mới
      description: |
        Endpoint mẫu để Provider nhận yêu cầu tạo cảnh báo từ Consumer.
        Sinh viên cần sửa theo cặp đàm phán thực tế.
      tags: [alerts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlertRequest'
            examples:
              unauthorizedAccess:
                value:
                  sourceService: core-business
                  alertType: UNAUTHORIZED_ACCESS
                  severity: HIGH
                  message: Phát hiện truy cập trái phép tại cổng chính
                  relatedEventId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
      responses:
        '201':
          description: Cảnh báo đã được tạo
          headers:
            Location:
              description: URI của alert vừa tạo.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
              examples:
                created:
                  value:
                    id: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
                    sourceService: core-business
                    alertType: UNAUTHORIZED_ACCESS
                    severity: HIGH
                    message: Phát hiện truy cập trái phép tại cổng chính
                    relatedEventId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
                    status: RESOLVED
                    createdAt: '2026-05-10T08:00:00Z'
                    resolvedAt: '2026-05-10T08:05:00Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      operationId: listAlerts
      summary: Lấy danh sách cảnh báo
      description: Lấy danh sách cảnh báo với cursor-based pagination.
      tags: [alerts]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Danh sách cảnh báo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertPage'
              examples:
                firstPage:
                  value:
                    items:
                      - id: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
                        sourceService: core-business
                        alertType: UNAUTHORIZED_ACCESS
                        severity: HIGH
                        message: Phát hiện truy cập trái phép tại cổng chính
                        relatedEventId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
                        status: RESOLVED
                        createdAt: '2026-05-10T08:00:00Z'
                        resolvedAt: '2026-05-10T08:05:00Z'
                    nextCursor: 'cursor-123'
                    hasMore: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /alerts/recent:
    get:
      operationId: getRecentAlerts
      summary: Lấy các cảnh báo gần đây
      description: Endpoint mẫu cho cặp Analytics ↔ Core Business hoặc Frontend ↔ Core Business.
      tags: [alerts]
      parameters:
        - name: limit
          in: query
          required: false
          description: Số cảnh báo tối đa cần lấy.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Danh sách cảnh báo gần đây
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    minItems: 0
                    maxItems: 100
                    items:
                      $ref: '#/components/schemas/Alert'
              examples:
                recentAlerts:
                  value:
                    items: []
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /alerts/{alertId}:
    get:
      operationId: getAlertById
      summary: Lấy chi tiết một cảnh báo
      description: Lấy chi tiết alert theo UUID.
      tags: [alerts]
      parameters:
        - $ref: '#/components/parameters/AlertId'
      responses:
        '200':
          description: Chi tiết cảnh báo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /events:
    post:
      operationId: createCampusEvent
      summary: Gửi sự kiện nghiệp vụ vào hệ thống
      description: |
        Endpoint minh họa polymorphism. Event có thể là SensorEvent hoặc AccessEvent.
      tags: [events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CampusEvent'
                - type: array
                  items:
                    $ref: '#/components/schemas/CampusEvent'
            examples:
              sensorEvent:
                value:
                  eventType: SENSOR_READING
                  eventId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
                  deviceId: SENSOR-001
                  metric: temperature
                  value: 38.5
                  unit: celsius
                  timestamp: '2026-05-10T08:00:00Z'
              accessEvent:
                value:
                  eventType: ACCESS_CHECK
                  eventId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babd
                  gateId: GATE-01
                  cardId: RFID-2026-001
                  decision: ALLOW
                  timestamp: '2026-05-10T08:01:00Z'
      responses:
        '201':
          description: Event đã được ghi nhận
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  /access/check:
    post:
      operationId: checkAccess
      summary: Yêu cầu mở cổng từ Access Gate
      description: B3 gửi yêu cầu mở cổng hoặc kiểm tra trạng thái thẻ tới Core Business.
      tags: [access]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Khóa idempotency để tránh xử lý quẹt thẻ trùng.
          schema:
            type: string
            minLength: 1
            maxLength: 100
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessCheckRequest'
      responses:
        '200':
          description: Core Business trả quyết định truy cập
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessCheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /vision/face-match:
    post:
      operationId: requestFaceMatch
      summary: Gửi yêu cầu nhận diện khuôn mặt tới AI Vision
      description: Core Business yêu cầu AI Vision so sánh ảnh và trả kết quả nhận diện.
      tags: [vision]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaceMatchRequest'
            examples:
              request:
                value:
                  imageRef: https://campus.local/images/card-scan-001.jpg
                  requestId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
                  cameraId: CAM-002
                  timestamp: '2026-05-10T08:01:00Z'
      responses:
        '201':
          description: Kết quả nhận diện khuôn mặt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaceMatchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

webhooks:
  alertCreated:
    post:
      operationId: onAlertCreated
      summary: Webhook khi có alert mới
      description: Provider phát sự kiện khi một cảnh báo mới được tạo.
      tags: [alerts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Alert'
      responses:
        '200':
          description: Consumer đã nhận webhook

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    AlertId:
      name: alertId
      in: path
      required: true
      description: UUID của cảnh báo.
      schema:
        type: string
        format: uuid
    Cursor:
      name: cursor
      in: query
      required: false
      description: Cursor dùng để lấy trang tiếp theo.
      schema:
        type: [string, 'null']
        minLength: 1
        maxLength: 200
    Limit:
      name: limit
      in: query
      required: false
      description: Số item tối đa mỗi trang.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  schemas:
    HealthStatus:
      type: object
      additionalProperties: false
      required: [status, service, time]
      properties:
        status:
          type: string
          enum: [ok]
          description: Trạng thái service.
          example: ok
        service:
          type: string
          minLength: 2
          maxLength: 80
          description: Tên service.
          example: lab02-contract-template
        time:
          type: string
          format: date-time
          description: Thời điểm kiểm tra.
          example: '2026-05-10T08:00:00Z'

    CreateAlertRequest:
      type: object
      required: [sourceService, alertType, severity, message]
      properties:
        sourceService:
          type: string
          minLength: 2
          maxLength: 80
          pattern: '^[a-z0-9-]+$'
          description: Service tạo alert.
          example: core-business
        alertType:
          type: string
          enum: [UNAUTHORIZED_ACCESS, SENSOR_THRESHOLD_EXCEEDED, UNKNOWN_PERSON, SYSTEM_ERROR]
          description: Loại cảnh báo.
          example: UNAUTHORIZED_ACCESS
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        message:
          type: string
          minLength: 5
          maxLength: 500
          description: Nội dung cảnh báo.
          example: Phát hiện truy cập trái phép tại cổng chính
        relatedEventId:
          type: [string, 'null']
          format: uuid
          description: Event liên quan, có thể null nếu chưa có event id.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc

    Alert:
      allOf:
        - $ref: '#/components/schemas/CreateAlertRequest'
        - type: object
          required: [id, status, createdAt, resolvedAt]
          properties:
            id:
              type: string
              format: uuid
              description: UUID của alert.
              example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
            status:
              type: string
              enum: [OPEN, ACKNOWLEDGED, RESOLVED]
              description: Trạng thái xử lý alert.
              example: OPEN
            createdAt:
              type: string
              format: date-time
              description: Thời điểm tạo alert.
              example: '2026-05-10T08:00:00Z'
            resolvedAt:
              type: [string, 'null']
              format: date-time
              description: Thời điểm xử lý xong, null nếu chưa xử lý.

    AlertSeverity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
      description: Mức độ nghiêm trọng của cảnh báo.
      example: HIGH

    AlertPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor, hasMore]
      properties:
        items:
          type: array
          minItems: 0
          maxItems: 100
          items:
            $ref: '#/components/schemas/Alert'
        nextCursor:
          type: [string, 'null']
          description: Cursor trang tiếp theo, null nếu hết dữ liệu.
        hasMore:
          type: boolean
          description: Còn dữ liệu ở trang sau hay không.
          example: false

    CampusEvent:
      oneOf:
        - $ref: '#/components/schemas/SensorEvent'
        - $ref: '#/components/schemas/AccessEvent'
      discriminator:
        propertyName: eventType
        mapping:
          SENSOR_READING: '#/components/schemas/SensorEvent'
          ACCESS_CHECK: '#/components/schemas/AccessEvent'

    SensorEvent:
      type: object
      additionalProperties: false
      required: [eventType, eventId, deviceId, metric, value, unit, timestamp]
      properties:
        eventType:
          type: string
          const: SENSOR_READING
          description: Loại event.
          example: SENSOR_READING
        eventId:
          type: string
          format: uuid
          description: UUID của event.
        deviceId:
          type: string
          pattern: '^SENSOR-[0-9]{3}$'
          description: Mã thiết bị cảm biến.
          example: SENSOR-001
        metric:
          type: string
          enum: [temperature, humidity, smoke, motion]
          description: Loại chỉ số cảm biến.
          example: temperature
        value:
          type: number
          minimum: -100
          maximum: 1000
          description: Giá trị đo được.
          example: 38.5
        unit:
          type: string
          minLength: 1
          maxLength: 20
          description: Đơn vị đo.
          example: celsius
        timestamp:
          type: string
          format: date-time
          description: Thời điểm đo.

    AccessEvent:
      type: object
      additionalProperties: false
      required: [eventType, eventId, gateId, cardId, decision, timestamp]
      properties:
        eventType:
          type: string
          const: ACCESS_CHECK
          description: Loại event.
          example: ACCESS_CHECK
        eventId:
          type: string
          format: uuid
          description: UUID của event.
        gateId:
          type: string
          pattern: '^GATE-[0-9]{2}$'
          description: Mã cổng.
          example: GATE-01
        cardId:
          type: string
          pattern: '^RFID-[0-9]{4}-[0-9]{3}$'
          description: Mã thẻ RFID.
          example: RFID-2026-001
        decision:
          type: string
          enum: [ALLOW, DENY]
          description: Quyết định truy cập.
          example: ALLOW
        timestamp:
          type: string
          format: date-time
          description: Thời điểm kiểm tra.

    AccessCheckRequest:
      type: object
      additionalProperties: false
      required: [cardId, gateId, timestamp]
      properties:
        cardId:
          type: string
          pattern: '^RFID-[0-9]{4}-[0-9]{3}$'
          description: Mã thẻ quẹt.
          example: RFID-2026-001
        gateId:
          type: string
          pattern: '^GATE-[0-9]{2}$'
          description: Mã cổng.
          example: GATE-01
        timestamp:
          type: string
          format: date-time
          description: Thời điểm quẹt thẻ.
          example: '2026-05-10T08:01:00Z'
        faceMatched:
          type: boolean
          description: Kết quả nhận diện khuôn mặt nếu có.
          example: true
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Độ tin cậy nhận diện khuôn mặt.
          example: 0.92

    AccessCheckResponse:
      type: object
      additionalProperties: false
      required: [decision, expiresAt]
      properties:
        decision:
          type: string
          enum: [ALLOW, DENY]
          description: Quyết định mở cổng.
          example: ALLOW
        expiresAt:
          type: string
          format: date-time
          description: Thời điểm quyết định hết hiệu lực.
          example: '2026-05-10T08:01:05Z'
        reasonCode:
          type: string
          description: Mã lý do từ Core Business.
          example: AUTHORIZED_CARD

    FaceMatchRequest:
      type: object
      additionalProperties: false
      required: [imageRef, requestId, cameraId, timestamp]
      properties:
        imageRef:
          type: string
          format: uri
          description: URL ảnh/nên upload trước.
          example: https://campus.local/images/card-scan-001.jpg
        requestId:
          type: string
          format: uuid
          description: UUID của request face match.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
        cameraId:
          type: string
          description: Mã camera.
          example: CAM-002
        timestamp:
          type: string
          format: date-time
          description: Thời gian chụp ảnh.

    FaceMatchResponse:
      type: object
      additionalProperties: false
      required: [detectionId, detectionType, faceMatched, isLive, confidence, status]
      properties:
        detectionId:
          type: string
          format: uuid
          description: ID của kết quả nhận diện.
        detectionType:
          type: string
          const: FACE
          description: Loại nhận diện.
        faceMatched:
          type: boolean
          description: Kết quả khớp khuôn mặt.
          example: true
        isLive:
          type: boolean
          description: Kết quả Liveness Detection chống giả mạo.
          example: true
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Độ tin cậy kết quả nhận diện.
          example: 0.92
        status:
          type: string
          enum: [success, low_confidence, no_face_detected]
          description: Trạng thái nhận diện.
          example: success
        matchedPersonId:
          type: [string, 'null']
          description: ID người được nhận diện nếu có.
          example: PERSON-1234

    EventAccepted:
      type: object
      additionalProperties: false
      required: [eventId, acceptedAt]
      properties:
        eventId:
          type: string
          format: uuid
          description: UUID của event được nhận.
        acceptedAt:
          type: string
          format: date-time
          description: Thời điểm hệ thống ghi nhận event.

    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: URI định danh loại lỗi.
          example: https://campus.local/errors/validation
        title:
          type: string
          minLength: 3
          maxLength: 120
          description: Tiêu đề lỗi ngắn gọn.
          example: Dữ liệu không hợp lệ
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: HTTP status code.
          example: 400
        detail:
          type: [string, 'null']
          maxLength: 500
          description: Mô tả chi tiết lỗi.
          example: Field cardId không đúng định dạng
        instance:
          type: [string, 'null']
          format: uri-reference
          description: URI của request gây lỗi.
          example: /alerts
        errors:
          type: array
          minItems: 0
          maxItems: 50
          items:
            type: object
            additionalProperties: false
            required: [field, code, message]
            properties:
              field:
                type: string
                minLength: 1
                maxLength: 80
                example: cardId
              code:
                type: string
                minLength: 2
                maxLength: 80
                example: PATTERN_MISMATCH
              message:
                type: string
                minLength: 3
                maxLength: 200
                example: cardId phải có dạng RFID-YYYY-NNN

  responses:
    BadRequest:
      description: Dữ liệu sai định dạng
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/validation
            title: Dữ liệu không hợp lệ
            status: 400
            detail: Payload không đúng JSON Schema
            instance: /alerts
            errors:
              - field: alertType
                code: ENUM_MISMATCH
                message: alertType không nằm trong danh sách cho phép

    Unauthorized:
      description: Thiếu hoặc sai thông tin xác thực
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/unauthorized
            title: Chưa xác thực
            status: 401
            detail: Thiếu Bearer token
            instance: /alerts/recent
            errors: []

    Forbidden:
      description: Không có quyền thực hiện thao tác
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

    NotFound:
      description: Resource không tồn tại
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

    Conflict:
      description: Xung đột nghiệp vụ hoặc trùng idempotency key
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

    UnprocessableEntity:
      description: Payload đúng định dạng nhưng không hợp lệ về nghiệp vụ
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

    InternalServerError:
      description: Lỗi không mong muốn phía server
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    TooManyRequests:
      description: Quá nhiều yêu cầu, thử lại sau
      headers:
        Retry-After:
          description: Số giây cần đợi trước khi gọi lại
          schema:
            type: integer
            minimum: 0
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'